Remove OAM DMA callback and minor cleanup
Removed some unnecessary lines of code and replaced "fmt" with "log"
No particular reason for it now, but could be useful in the future.

A wrapper has been created for the cpu's read and write instead of
calling the read and write function from the cpu bus directly. This
way, the cpu can check itself if it is writing to the DMA register and
increment the cycles to skip directly instead of the bus struct having
to call the cpu. This also gives a single location for any logging or
debug logic to go for cpu reads and writes. Honestly, this should have
been done from the start.