Add UxROM mapper support
I tested the changes with Mega Man and it appears to work just fine. I
also decided to stick with just .NES 1.0 headers and not to support
trainers for now. As far as I can tell, the majority of ROMs for
official Nintendo games all use the early header version, and those are
what I'm looking to support. Overall, the mapper creation and ROM
parsing has been cleaned up a bit.